Add gson mapping for repository implementation.
Motivation
----------
A mapping that only supports some primitive data types is very restricted and useless when
you have real models with relations and abstract data structures.

Modifications
-------------
Added son entity converter and mapping mode setting that is also implemented in the corresponding repository and bucket interface.

Result
------
A powerful gson implementation with the main disadvantage that the @Id and @Field annotations aren't supported anymore. The id field has to be named "id" (take a look at @SerializedName for field naming). In a future release we can provide a custom son naming policy based on the @Id and @Field annotations.

Change-Id: I71f5ac50a15b2383b33b240904dd2b1e3d8f5694